Run ef-tests
