+++ crates/evm/src/ef_tests/cases/blockchain_test.rs
@@ -5,25 +5,22 @@ use std::fs;
 use std::path::Path;
 use std::sync::Arc;
 
-use alloy_rlp::Decodable;
+use alloy_rlp::{Decodable, Encodable};
 use rayon::iter::{ParallelBridge, ParallelIterator};
 use reth_db::test_utils::{create_test_rw_db, create_test_static_files_dir};
-use reth_primitives::{BlockBody, SealedBlock, StaticFileSegment, TransactionSignedEcRecovered};
+use reth_primitives::{BlockBody, SealedBlock, StaticFileSegment};
 use reth_provider::providers::StaticFileWriter;
 use reth_provider::{BlockReader, HashingWriter, ProviderFactory};
-use reth_revm::primitives::CfgEnvWithHandlerCfg;
+use sov_modules_api::default_context::DefaultContext;
+use sov_modules_api::utils::generate_address;
+use sov_modules_api::{Context, WorkingSet};
+
 // use reth_stages::stages::ExecutionStage;
 // use reth_stages::{ExecInput, Stage};
-use sov_modules_api::WorkingSet;
-use sov_prover_storage_manager::new_orphan_storage;
-
 use crate::ef_tests::models::{BlockchainTest, ForkSpec};
 use crate::ef_tests::{Case, Error, Suite};
-use crate::evm::db::EvmDb;
-use crate::evm::executor::execute_multiple_tx;
-use crate::evm::handler::CitreaExternal;
-use crate::evm::primitive_types::BlockEnv;
-use crate::{get_cfg_env, EvmChainConfig};
+use crate::test_utils::{commit, get_evm_with_storage};
+use crate::{EvmConfig, RlpEvmTransaction};
 
 /// A handler for the blockchain test suite.
 #[derive(Debug)]
@@ -152,33 +149,10 @@ impl Case for BlockchainTestCase {
                     .commit_without_sync_all()
                     .unwrap();
 
-                // Execute the execution stage using the EVM processor factory for the test case
-                // network.
-                // let _ = ExecutionStage::new_with_factory(EvmProcessorFactory::new(
-                //     Arc::new(case.network.clone().into()),
-                //     EvmConfig::default(),
-                // ))
-                // .execute(
-                //     &provider,
-                //     ExecInput {
-                //         target: last_block.as_ref().map(|b| b.number),
-                //         checkpoint: None,
-                //     },
-                // );
+                let (evm, working_set, storage) = get_evm_with_storage(&EvmConfig::default());
+                let root = commit(working_set, storage.clone());
+                let mut working_set: WorkingSet<DefaultContext> = WorkingSet::new(storage.clone());
 
-                let mut working_set = WorkingSet::new(new_orphan_storage(db.path()).unwrap());
-                let block_env = BlockEnv {
-                    gas_limit: reth_primitives::constants::ETHEREUM_BLOCK_GAS_LIMIT.into(),
-                    ..Default::default()
-                };
-                let cfg = EvmChainConfig::default();
-                let cfg_env: CfgEnvWithHandlerCfg = get_cfg_env(&block_env, cfg, None);
-                let mut citrea_handler_ext = CitreaExternal::new(0);
-
-                type C = sov_modules_api::default_context::DefaultContext;
-                let evm = crate::Evm::<C>::default();
-                let evm_db: EvmDb<'_, C> = evm.get_db(&mut working_set);
-                let mut cumulative_gas_used = 0;
                 let block = provider
                     .block_with_senders(
                         reth_primitives::BlockHashOrNumber::Number(
@@ -186,26 +160,40 @@ impl Case for BlockchainTestCase {
                         ),
                         reth_provider::TransactionVariant::NoHash,
                     )
-                    .unwrap()
+                    .transpose()
+                    .map(|r| r.ok())
+                    .flatten()
                     .unwrap();
 
-                let txs: Vec<TransactionSignedEcRecovered> =
-                    block.into_transactions_ecrecovered().collect();
+                let txs: Vec<RlpEvmTransaction> = block
+                    .transactions()
+                    .map(|t| {
+                        let mut buffer = Vec::<u8>::new();
+                        t.encode(&mut buffer);
+                        RlpEvmTransaction { rlp: buffer }
+                    })
+                    .collect();
 
                 // Call begin_soft_confirmation_hook
-                // evm.execute_call or something similar
-                let _ = execute_multiple_tx(
-                    evm_db,
-                    block_env,
-                    &txs,
-                    cfg_env,
-                    &mut citrea_handler_ext,
-                    cumulative_gas_used,
+                evm.begin_soft_confirmation_hook(
+                    [0u8; 32],
+                    0,
+                    [0u8; 32],
+                    &root,
+                    0,
+                    0,
+                    &mut working_set,
                 );
 
-                // end_soft_confirmation_hook
-                // commit
-                // finalize_hook
+                let sender_address = generate_address::<DefaultContext>("sender");
+                let sequencer_address = generate_address::<DefaultContext>("sequencer");
+                let context = DefaultContext::new(sender_address, sequencer_address, 1);
+                let _ = evm.execute_call(txs, &context, &mut working_set);
+
+                evm.end_soft_confirmation_hook(&mut working_set);
+                let root = commit(working_set, storage.clone());
+                let mut working_set: WorkingSet<DefaultContext> = WorkingSet::new(storage.clone());
+                evm.finalize_hook(&root.into(), &mut working_set.accessory_state());
 
                 // Validate the post-state for the test case.
                 match (&case.post_state, &case.post_state_hash) {
