+++ crates/evm/src/tests/call_tests.rs
@@ -13,7 +13,7 @@ use crate::evm::primitive_types::Receipt;
 use crate::smart_contracts::{
     BlockHashContract, LogsContract, SelfDestructorContract, SimpleStorageContract, TestContract,
 };
-use crate::tests::genesis_tests::get_evm;
+use crate::test_utils::get_evm;
 use crate::tests::test_signer::TestSigner;
 use crate::{AccountData, EvmConfig, RlpEvmTransaction, DEFAULT_CHAIN_ID};
