+++ crates/evm/src/tests/genesis_tests.rs
@@ -8,17 +8,12 @@ use reth_primitives::{
     Address, BaseFeeParams, Bloom, Bytes, Header, SealedHeader, B256, EMPTY_OMMER_ROOT_HASH,
 };
 use revm::primitives::{SpecId, KECCAK_EMPTY, U256};
-use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::prelude::*;
-use sov_modules_api::{Module, WorkingSet};
-use sov_prover_storage_manager::new_orphan_storage;
 
-use super::queries::commit;
 use crate::evm::primitive_types::SealedBlock;
 use crate::evm::{AccountInfo, DbAccount, EvmChainConfig};
-use crate::{AccountData, Evm, EvmConfig};
-
-type C = DefaultContext;
+use crate::test_utils::{get_evm, GENESIS_HASH, GENESIS_STATE_ROOT};
+use crate::{AccountData, EvmConfig};
 
 lazy_static! {
     pub(crate) static ref TEST_CONFIG: EvmConfig = EvmConfig {
@@ -61,12 +56,6 @@ lazy_static! {
         base_fee_params: BaseFeeParams::ethereum(),
     };
 
-    pub(crate) static ref GENESIS_HASH: B256 = B256::from(hex!(
-        "d3d0598f798aeac0eb991a29a1ec2e79e6d3caf76393c787a564d89233342df8"
-    ));
-    pub(crate) static ref GENESIS_STATE_ROOT: B256 = B256::from(hex!(
-        "050d41ea04851c5800ba402d42d7a3777ca025feda7c1bd6a272ff9710a4501a"
-    ));
     pub(crate) static ref GENESIS_DA_TXS_COMMITMENT: B256 = B256::from(hex!(
         "4a5e1e4baab89f3a32518a88c31bc87f618f76673e2cc77ab2127b7afdeda33b"
     ));
@@ -277,29 +266,4 @@ fn genesis_head() {
     assert_eq!(genesis_block.l1_fee_rate, 0);
 
     assert_eq!(genesis_block.transactions, (0u64..0u64));
-}
-
-pub(crate) fn get_evm(config: &EvmConfig) -> (Evm<C>, WorkingSet<DefaultContext>) {
-    let tmpdir = tempfile::tempdir().unwrap();
-    let storage = new_orphan_storage(tmpdir.path()).unwrap();
-    let mut working_set = WorkingSet::new(storage.clone());
-    let evm = Evm::<C>::default();
-    evm.genesis(config, &mut working_set).unwrap();
-
-    let root = commit(working_set, storage.clone());
-
-    let mut working_set: WorkingSet<DefaultContext> = WorkingSet::new(storage.clone());
-    evm.finalize_hook(&root.into(), &mut working_set.accessory_state());
-
-    evm.begin_soft_confirmation_hook([1u8; 32], 1, [2u8; 32], &root, 0, 0, &mut working_set);
-    evm.end_soft_confirmation_hook(&mut working_set);
-
-    let root = commit(working_set, storage.clone());
-    let mut working_set: WorkingSet<DefaultContext> = WorkingSet::new(storage.clone());
-    evm.finalize_hook(&root.into(), &mut working_set.accessory_state());
-
-    // let mut genesis_state_root = [0u8; 32];
-    // genesis_state_root.copy_from_slice(GENESIS_STATE_ROOT.as_ref());
-
-    (evm, working_set)
 }
