+++ crates/evm/src/tests/hooks_tests.rs
@@ -7,11 +7,12 @@ use reth_primitives::{
 };
 use sov_modules_api::{StateMapAccessor, StateValueAccessor, StateVecAccessor};
 
-use super::genesis_tests::{get_evm, GENESIS_DA_TXS_COMMITMENT, TEST_CONFIG};
+use super::genesis_tests::{GENESIS_DA_TXS_COMMITMENT, TEST_CONFIG};
 use crate::evm::primitive_types::{
     Block, BlockEnv, Receipt, SealedBlock, TransactionSignedAndRecovered,
 };
-use crate::tests::genesis_tests::{BENEFICIARY, GENESIS_STATE_ROOT};
+use crate::test_utils::{get_evm, GENESIS_STATE_ROOT};
+use crate::tests::genesis_tests::BENEFICIARY;
 use crate::{PendingTransaction, DEFAULT_CHAIN_ID};
 
 lazy_static! {
