+++ crates/evm/src/tests/queries/mod.rs
@@ -10,17 +10,15 @@ use revm::primitives::{SpecId, KECCAK_EMPTY, U256};
 use sov_modules_api::default_context::DefaultContext;
 use sov_modules_api::utils::generate_address;
 use sov_modules_api::{Context, Module, WorkingSet};
-use sov_prover_storage_manager::{new_orphan_storage, SnapshotManager};
-use sov_state::{DefaultStorageSpec, ProverStorage, Storage};
 
 use crate::call::CallMessage;
 use crate::smart_contracts::{
     CallerContract, LogsContract, SimplePayableContract, SimpleStorageContract,
 };
+use crate::test_utils::{commit, get_evm_with_storage};
 use crate::tests::call_tests::{
     create_contract_transaction, publish_event_message, set_arg_message,
 };
-use crate::tests::genesis_tests::GENESIS_STATE_ROOT;
 use crate::tests::test_signer::TestSigner;
 use crate::{AccountData, Evm, EvmConfig, RlpEvmTransaction};
 
@@ -312,49 +310,4 @@ pub fn init_evm_with_caller_contract() -> (Evm<C>, WorkingSet<C>, TestSigner) {
     let working_set: WorkingSet<DefaultContext> = WorkingSet::new(prover_storage);
 
     (evm, working_set, dev_signer)
-}
-
-fn get_evm_with_storage(
-    config: &EvmConfig,
-) -> (
-    Evm<C>,
-    WorkingSet<DefaultContext>,
-    ProverStorage<DefaultStorageSpec, SnapshotManager>,
-) {
-    let tmpdir = tempfile::tempdir().unwrap();
-    let prover_storage = new_orphan_storage(tmpdir.path()).unwrap();
-    let mut working_set = WorkingSet::new(prover_storage.clone());
-    let evm = Evm::<C>::default();
-    evm.genesis(config, &mut working_set).unwrap();
-
-    let mut genesis_state_root = [0u8; 32];
-    genesis_state_root.copy_from_slice(GENESIS_STATE_ROOT.as_ref());
-
-    evm.finalize_hook(
-        &genesis_state_root.into(),
-        &mut working_set.accessory_state(),
-    );
-    (evm, working_set, prover_storage)
-}
-
-pub(crate) fn commit(
-    working_set: WorkingSet<DefaultContext>,
-    storage: ProverStorage<DefaultStorageSpec, SnapshotManager>,
-) -> [u8; 32] {
-    // Save checkpoint
-    let mut checkpoint = working_set.checkpoint();
-
-    let (cache_log, witness) = checkpoint.freeze();
-
-    let (root, authenticated_node_batch) = storage
-        .compute_state_update(cache_log, &witness)
-        .expect("jellyfish merkle tree update must succeed");
-
-    let working_set = checkpoint.to_revertable();
-
-    let accessory_log = working_set.checkpoint().freeze_non_provable();
-
-    storage.commit(&authenticated_node_batch, &accessory_log);
-
-    root.0
 }
